solution refactoring into tokenized
